Fix snippets, update code
