+++ docs/csharp/programming-guide/classes-and-structs/snippets/access-modifiers/Program.cs
@@ -1,0 +1,1 @@
+﻿NestedTypes.Example();
