+++ docs/csharp/programming-guide/classes-and-structs/snippets/access-modifiers/nestedtypes.cs
