--- samples/snippets/csharp/objectoriented/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace objectoriented
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            NestedTypes.Example();
-        }
-    }
-}
